Arreglar

+++ formato-factura/models/linea_servicio.py
@@ -11,7 +11,7 @@ class LineaServicio( models.Model):
     move_id = fields.Many2one( comodel_name = "account.move", string = "Linea de Servicio")
     nombre_cargo = fields.Char( string = "Cargo")
     cantidad = fields.Integer( string = "Cantidad")
-    precio_unidad = fields.Integer( string = "Precio")
-    subtotal = fields.float( string = "Subtotal")
+    precio_unidad = fields.Float( string = "Precio")
+    subtotal = fields.Float( string = "Subtotal")
